RavenDB-16499 Moved to dedicated namespace
